requirements.txt
absl-py==1.2.0
anyio==4.3.0
apache-beam==2.22.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
Automat==0.8.0
avro-python3==1.9.2.1
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
bokeh==3.1.1
cachetools==5.3.3
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
cloud-init==23.4.4
cloudpickle==2.2.1
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cramjam==2.8.3
crcmod==1.7
cryptography==2.8
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
dm-sonnet==2.0.0
dm-tree==0.1.8
dnspython==2.6.1
docopt==0.6.2
entrypoints==0.3
etils==1.3.0
exceptiongroup==1.2.0
executing==2.0.1
facets-overview==1.1.1
fastavro==0.23.6
fasteners==0.19
fastjsonschema==2.19.1
flatbuffers==1.12
fonttools==4.50.0
future==0.18.3
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
hdfs==2.7.3
httpcore==1.0.4
httplib2==0.14.0
httpx==0.27.0
hyperlink==19.0.0
idna==2.8
imageio==2.20.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
incremental==16.10.1
intervaltree==3.1.0
ipykernel==6.14.0
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
Js2Py==0.74
json5==0.9.24
jsonpatch==1.22
jsonpickle==3.0.3
jsonpointer==2.0
jsonschema==4.18.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.1
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter-server==2.13.0
jupyter-server-terminals==0.5.3
jupyterlab==4.1.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.4
jupyterlab-widgets==3.0.10
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
librosa==0.7.2
llvmlite==0.32.1
magenta==2.1.4
Markdown==3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mido==1.2.6
mir-eval==0.7
mistune==3.0.2
mock==2.0.0
more-itertools==4.2.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
note-seq==0.0.3
notebook==7.1.2
notebook-shim==0.2.4
numba==0.49.1
numpy==1.21.6
oauth2client==3.0.0
oauthlib==3.1.0
objsize==0.7.0
opt-einsum==3.3.0
orjson==3.9.15
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pbr==6.0.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.5.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
pretty-midi==0.2.9
prometheus-client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==0.17.1
pyarrow-hotfix==0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydot==1.4.2
pydub==0.25.1
pygments==2.17.2
PyGObject==3.36.0
pygtrie==2.5.0
PyHamcrest==1.9.0
pyjsparser==2.7.1
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-json-logger==2.0.7
python-rtmidi==1.1.2
python-snappy==0.7.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
resampy==0.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.7.3
SecretStorage==2.3.1
Send2Trash==1.8.2
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sk-video==1.1.10
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.5.6
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
ssh-import-id==5.10
stack-data==0.6.3
systemd-python==234
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.13.0
tensorflow-probability==0.17.0
termcolor==2.4.0
terminado==0.18.1
testresources==2.0.1
tf-slim==1.1.0
threadpoolctl==3.4.0
tifffile==2023.7.10
timeloop==1.0.2
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
Twisted==18.9.0
typing-extensions==3.7.4.3
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
xyzservices==2023.10.1
yarg==0.1.9
zipp==3.18.1
zope.interface==4.7.1
zstandard==0.22.0
